feat: Half-life 2.
Opens main menu, crashes after spawning the player.

+++ game/server/basecombatweapon.cpp
@@ -39,15 +39,15 @@ extern int	gEvilImpulse101;		// In Player.h
 // -----------------------------------------
 //	Sprite Index info
 // -----------------------------------------
-int		g_sModelIndexLaser;			// holds the index for the laser beam
+short		g_sModelIndexLaser;			// holds the index for the laser beam
 const char	*g_pModelNameLaser = "sprites/laserbeam.vmt";
-int		g_sModelIndexLaserDot;		// holds the index for the laser beam dot
-int		g_sModelIndexFireball;		// holds the index for the fireball
-int		g_sModelIndexSmoke;			// holds the index for the smoke cloud
-int		g_sModelIndexWExplosion;	// holds the index for the underwater explosion
-int		g_sModelIndexBubbles;		// holds the index for the bubbles model
-int		g_sModelIndexBloodDrop;		// holds the sprite index for the initial blood
-int		g_sModelIndexBloodSpray;	// holds the sprite index for splattered blood
+short		g_sModelIndexLaserDot;		// holds the index for the laser beam dot
+short		g_sModelIndexFireball;		// holds the index for the fireball
+short		g_sModelIndexSmoke;			// holds the index for the smoke cloud
+short		g_sModelIndexWExplosion;	// holds the index for the underwater explosion
+short		g_sModelIndexBubbles;		// holds the index for the bubbles model
+short		g_sModelIndexBloodDrop;		// holds the sprite index for the initial blood
+short		g_sModelIndexBloodSpray;	// holds the sprite index for splattered blood
 
 
 ConVar weapon_showproficiency( "weapon_showproficiency", "0" );
